Makefile -- updated to reflect the new files marked_abeliangroup.cpp, marked_abeliangrou p.h, homologicaldata.h and homologicaldata.cpp
marked_abeliangroup.h -- this gives the markedAbelianGroup and homMarkedAbelianGroup group classes. markedAbelianGroup is a `souped-up' NAbelianGroup class, that allows to keep track of elements in the group. This is needed for the homMarkedAbelianGroup class.  We will want to keep NAbelianGroup though since the algorithms in NAbelianGroup are likely a bit faster than these. Might want to test that. homMarkedAbelianGroup keeps track of
homomorphisms between markedAbelianGroups and gives various invariants: kernel, cokernel
, image. We still need to make python wrappers for these classes.

marked_abeliangroup.cpp -- corresponding .cpp file to marked_abeliangroup.h

homologicaldata.h -- this is a `wrapper' class for all the various homological invariant
s of a triangulated 3-manifold.  Right now it gives: the homology groups in two differen
t (transverse) coordinate systems, plus the natural isomorphism between the two on the l
evel of H0 and H1 (I have not implemented the isomorphism for H2 yet) given by a cellular approximation to the identity from the one CW-decomposition to the other.  This allows for the computation of the H1 torsion linking form, which we also construct the Kawauchi-Kojima complete set of invariants for, and the various obstructions to embedding in an oriented homology 4-sphere. We need a python wrapper for this class.  From this data it won't be difficult to implement the cohomology ring, at least in the form of a presentation w/generators and relations. At present I don't see that being very useful -- maybe GAP or something
similar would have algorithms to construct useful invariants of a ring given via a
multiplication table? 

homologicaldata.cpp -- supporting .cpp file to homologicaldata.h.

ntriangulation.h -- made homologicalData a friend.

ntetrahedron.h -- made homologicalData a friend.