nmarkedabeliangroup and nhommarkedabelian group have recieved several significant changes:
nmarkedabeliangroup changes:

a new constructor for nmarkedabeliangroups, to easily construct n Z_p 
where p is an integer -- when p==0 this is interpreted as Z. 

changed == for nmarkedabeliangroups to demand that the groups have the 
same chain complex definition.  the old == has been turned into 
isIsomorphicTo()

added getCCRep() to nmarkedabeliangroup -- this converts from SNF to the 
chain complex coordinates.  

added isCycle() to determine if an element of the chain complex is a 
cycle. 

added some todo items. 

nhommarkedabeliangroup changes:

add a new constructor to nhommarkedabelian group, to allow for the 
defining matrix to be in the SNF coordinates of the respective groups, 
rather than in the chain complex coordinates. 

add isIdentity() query to nhommarked -- checks that the domain and the 
range are identical (==) and if the induced map on homology is the 
identity map. 

add evalCC and evalSNF -- to compute the image of an element in the 
group under the homomorphism, in either chain complex or SNF 
coordinates. 

add inverseHom() routine which returns the inverse of the homomorphism 
if it is an isomorphism, and the 0 homomorphism otherwise. 

add a composition of homomorphism operation *

small doc tweaks all over, including nbinaryS4.h and nbinaryA5.h