Added bdryMap to allow one to see what the differential is when applied to an element of the chain complex.
Added isBoundary() to determine if your element of the chain complex is 
a boundary or not. 

Added writeAsBoundary to find an element whose differential is this 
element.  I think I haven't implemented this one yet but it'll come 
soon.  I only get a few chances to svn up nowadays so I'll take what I 
can get. 

isChainMap determines if two consecutive NHomMarkedAbelianGroups make a 
chain map. 

renamed old isChainMap to isCycleMap to reflect the old inaccuracy in 
terminology.