Added a routine to compute the H2-intersection product for a 3-manifold. At present it's slightly crippled by our smithNormalForm routine as there's the occasional 3-manifold triangulation where the change-of-basis matrices will get totally out of control.  I'll keep working on that but there's no harm in svn upping this stuff now.
Anyhow, to support this there's a new routine computeBaryCC() which
computes the chain complex for the barycentric subdivision, and the
chain maps between this and the standard and dual homology chain
complexes.  Since the barycentric subdivision is so fat I figured it's
best to keep this totally seperate.

getMixedHomology(q) is the requisite call for the homology in
barycentric subdivision coordinates, and getStandardToMixedHom and
getDualToMixedhom are the NHomMarkedAbelianGroups relating the two.

imgH2form is the current routine that supports the h2-intersection form
but its name and implementation are still subject to quite a bit of flux
at the moment.  In particular, I'll eventually get the mod-p
coefficients computation up and running for this.  Should be another
week or so on that.

One of the new routines is verifyChainComplexes and
verifyCoordinateIsomorphisms() which checks that everything that should
be a chain map and/or isomorphism and/or identity map really is.  I
think one of these needs to be tweaked once I get my code using Ben's
orientations for the link of a vertex and syncing that with all the
chain complexes and maps.