Beginning migration to new more flexible chain complex types.
Prototypes for the new chain complexes are in ncellulardata, this has required a splitting of NSparseGrid into a base class and an inherited class.  NSparseGridRing is the inherited class which assumes the template class T is a ring-object, ie suitable for linear algebra.   NSparseGrid makes less assumptions on T. 

All the other updates are derivative tweaks of the above changes.