Add first revision of ReginaManager class
This is part of the removal of KParts. To allow multiple windows,
there needs to be a parent, or managing class (hence the name).
This class simply keeps a track of which windows have been created.
The windows will be of the class ReginaMain, so everything should work
fine for now. Drag+drop of documents between windows needs to be 
implemented manually, but otherwise things should "work".