Modified proliferateRelators() so that it not only constructs a*b where a and be are cyclic permutations of relators, but also a*b^-1.   Also added some efficiencies to avoid adding some such pairs (if it's known intelligentSimplify() will quickly erase them).
In effect this means applying proliferateRelators then intelligentSimplify() is more effective than previously, but it also results in a slightly more aggressive exponential growth.  So the 1st attempt is more effective but the 2nd (iterative) application can be slower. 

When run against the 25 homotopy 4-spheres intelligentSimplify couldn't reduce the pi1 presentations of, now 24 are handled on the 1st proliferateRelators intelligentSimplify() cycle, and only one requires proliferateRelators twice.