Cycle drawing
Adds another tab to the skeleton tab, called Cycle Drawing. This draws out the
decomposition representation of the manifold. It also allows manipulation of
the drawing (by moving edges around) but you cannot change the underlying
triangulation through it.

This still has a few issues. Firstly, we still need some method of swapping
faces around to make nicer pictures. Also, the lines tend to overlap if there
is symmetry, I don't yet have a nice solution for that.