Code to create enumeration databases
This creates databases for used when enumerating triangulations. The file
format required as input is exactly the same as for mkcensusdb, that is, a
(possibly gzipped) text file containing isomorphism signatures and names of
corresponding triangulations, i.e.:
<isosig> <name>
<isosig> <name>
...

Note that the name is not used by mkenumerationdb, but to maintain
compatability with mkcensusdb the code does look for it.