Add enumeration database
This is a database that can be used when enumerating triangulations. The
database itself stores a list of "interesting" triangulations, sorted by
underlying face pairing graph. The class is used to look up such entries. Note
that the appropriate database will need to be constructed with mkenumerationdb,
in engine/data/census. Also note that Regina does not in any way validate that
the correct enumeration database is being used for a census. Using an
enumeration database that does not match the current census settings (i.e.
orientability, minimality conditions, boundary conditions) will result in an
incomplete census.