Add TetIntChooser and 0-2 move to Qt GUI
TetIntChooser just lets the user choose a tetrahedron and integer.
Note that it displays slightly differently to other choosers, it has a
separate combo box for tetrahedron and integer.
This is because I'm using it for 0-2 moves, which are always valid. If I
tried to combine the two, then a triangulation on 4 tetrahedra would
already have 24 options for tetrahedron+edge.