Adds overrided __or__ to flags
This somewhat addresses #51 by creating an __or__ operator for each of
the enums to promote the left hand side to NormalList/NormalAlg
respectively. This lines up with the engine, where only the or operator
promotes. Note that using pybind11::self | pybind11::self does not
compile (not sure on exact reason), and the .def has to come after
export_values(), as otherwise it demotes the pybind11::enum down to a
pybind11::class which means it loses access to export_values().

Not sure if this is the correct way forward still, but it's one way
forward.