Remove "Hyperbolic" from name of example data set
The "Hyperbolic knot / link census" data set includes at least one
complement of a non-hyperbolic manifold (e.g., the complement of the
graph manifold L310066). This changes the filenames to remove the "hyp-"
prefix, changes the displayed name of the examples to remove the
word Hyperbolic, and updates the CensusDB lookup to use the new
filename.

Fixes issue #45

Signed-off-by: William Pettersson <william@ewpettersson.se>